requirements.txt
pytest
appium-python-client
selenium
allure-pytest